Fix `next` by clearing caches after handling syscall breakpoints.
Broke when fixed network recv handling - only realized now.

It's possible but unlikely there are other places we refresh the caches when we
shouldn't - only one I'm aware of is threadStacktrace/ThreadStacktrace,
which is only called from handleSyscallBreakpoints and
taintedSyscallEntry (but latter isn't causing problem).

Need to understand when cache is cleared/refreshed a bit better to know exactly
when refreshing cache has unwanted effects (see logging and test target
on broken-next branch). Overview:

- GetG sets thread.Common().g if it's nil, else just returns it
- ClearCaches clears thread.Common().g

- In the last resume before hitting line 11 bp, handleSyscallEntry sets
  thread.Common().g to the g at syscall entry (via threadStacktrace)
- On next from line 11 bp, uses stale g at syscall entry to set condition for
  line 12 bp (in particular g.stackHi)
- On hit line 12 bp, refreshes g to line 12 before checking condition -
  thus condition doesn't pass (expects stackHi to be unchanged from
when bp was set)