Implement config file read breakpoints.
Dlv is configured with list of target's config files (passes them to
Target config, via Debugger's config) - when one is read,
returns syscall entry bp to client which watches the read buffer. When
that watchpoint hits, fill in mem-param entries for that portion of
buffer with contents of buffer (if they're not already filled).

Also move logic to get syscall info (from registers or /proc/) to
server, which passes result to client via syscall entry bp.

And change client and proc tests to only build target with forked go, so
we can use whatever go version we want for client and dlv.