Debugging failure to set wp on s[len(s)-1] for
wp hit on parser.buffer[buffer_len+0], in scannerc.go 525 (read()),
when running with control flow, only for the second two hits (first is
ok). Due to evalWatchexpr saying parser is nil.

This commit has extra logging in read(), causing a different variable
(s) to be incorrectly nil. Makes it crash instead of fail to set wp, so
will likely go back to other case where parser is nil, but this commit
shows that logging changes eval behavior. Don't think it's a timing
issue since there's no concurrency - suspect refreshing a memory
cache...
(see tmp.md - the hits, in this case 2 but 3 without the extra logging, are marked with "AAH")