Add a new behavior type: function args.
Also add a test that outputs a graph for the caddy manual scan, with
only the http response for now (i.e. not using the new behavior type yet).
Has a node for the body param, with edges to nodes for each of the 13 offsets in the response body.

Graph renders with ConfTamer-query
(before new type was added, which requires updating that repo).